Moved image service into the feature folder i.e. image-upload.
